+++ packages/@expo/cli/src/utils/FileNotifier.ts
@@ -33,11 +33,11 @@ export class FileNotifier {
     return null;
   }
 
-  public startObserving() {
+  public startObserving(callback?: (cur: any, prev: any) => void) {
     const configPath = this.resolveFilePath();
     if (configPath) {
       debug(`Observing ${configPath}`);
-      return this.watchFile(configPath);
+      return this.watchFile(configPath, callback);
     }
     return configPath;
   }
@@ -49,7 +49,7 @@ export class FileNotifier {
   /** Watch the file and warn to reload the CLI if it changes. */
   public watchFile = memoize(this.startWatchingFile.bind(this));
 
-  private startWatchingFile(filePath: string): string {
+  private startWatchingFile(filePath: string, callback?: (cur: any, prev: any) => void): string {
     const configName = path.relative(this.projectRoot, filePath);
     const listener = (cur: any, prev: any) => {
       if (prev.size || cur.size) {
@@ -61,7 +61,7 @@ export class FileNotifier {
       }
     };
 
-    const watcher = watchFile(filePath, listener);
+    const watcher = watchFile(filePath, callback ?? listener);
 
     this.unsubscribe = () => {
       watcher.unref();